Big refactor to make code live in the correct file / a saner place and adding a great deal of jsdoc
