+++ jsdoc.conf
@@ -1,7 +1,8 @@
 {
     "source": {
         "include": [
-          "./src"
+          "./src",
+          "./README.md"
         ]
     },
     "opts": {
