+++ package.json
@@ -2,12 +2,12 @@
   "name": "sw-testing-helpers",
   "version": "0.0.13",
   "description": "A set of helper files used to test Propel and sw-toolbox.",
-  "main": "./src/node/index.js",
+  "main": "./index.js",
   "scripts": {
     "build": "gulp",
     "build-docs": "jsdoc -c ./jsdoc.conf",
     "bundle": "./project/copy-build-files.sh",
-    "test": "mocha"
+    "test": "gulp scripts:lint && mocha"
   },
   "repository": {
     "type": "git",
@@ -32,7 +32,7 @@
     "babel-preset-es2015": "^6.6.0",
     "babelify": "^7.2.0",
     "browserify": "^13.0.0",
-    "eslint": "^2.4.0",
+    "eslint": "^2.8.0",
     "eslint-config-google": "^0.5.0",
     "glob": "^7.0.3",
     "gulp": "^3.9.1",
@@ -42,7 +42,6 @@
     "gulp-uglify": "^1.5.3",
     "jsdoc": "^3.4.0",
     "node-fetch": "^1.5.0",
-    "pullr": "^0.3.2",
     "require-dir": "^0.3.0",
     "rollupify": "^0.2.0",
     "run-sequence": "^1.1.5",
